requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pdf2image==1.16.3
python-docx==1.1.0
Pillow==10.1.0
mistralai==0.4.2
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.6
tiktoken==0.5.2